+++ examples/albert/deepspeed_hf.py
@@ -14,7 +14,7 @@ from slapo.logger import get_logger
 from slapo.op.cross_entropy import ParallelCrossEntropy
 from slapo.utils.report import report_memory
 
-from model import schedule_model
+from slapo.model_schedule.albert import schedule_model
 from examples.utils import (
     train_with_deepspeed_engine,
     get_ds_config,
